+++ __tests__/__mocks__/request.js
@@ -1,6 +1,19 @@
-export default function request(obj, cb) {
+let request = (obj, cb) => {
   process.nextTick(() => {
-    const err = obj.url.match(/error/i) ? true : false;
-    cb(err, { statusCode: 200 }, {})
+    cb(request.__err, request.__resp, request.__body);
   })
 }
+
+request.__setErr = (err) => {
+  request.__err = err;
+}
+
+request.__setResp = (resp) => {
+  request.__resp = resp;
+}
+
+request.__setBody = (body) => {
+  request.__body = body;
+}
+
+export default request;