+++ src/events/CognitoStream/constants.js
@@ -1,0 +1,3 @@
+export const LOCAL = 'Inline';
+ 
+export const REMOTE = 'S3Url';