+++ src/events/DynamoDB/record.js
@@ -1,5 +1,6 @@
+import { unmarshalItem } from 'dynamodb-marshaler';
+
 import { CREATE, UPDATE, DELETE } from './constants';
-import Document from './document';
 
 export default class {
   constructor(record) {
@@ -7,7 +8,7 @@ export default class {
   }
 
   get type() {
-    switch (this.record.StreamViewType) {
+    switch (this.record.eventName) {
       case 'INSERT':
         return CREATE;
       case 'MODIFY':
@@ -15,20 +16,20 @@ export default class {
       case 'REMOVE':
         return DELETE;
       default:
-        return undefined;
+        return this.record.eventName.toLowerCase();
     }
   }
 
   get keys() {
-    return this.data.Keys;
+    return unmarshalItem(this.data.Keys);
   }
 
-  get newDocument() {
-    return new Document(this.data.NewDocument);
+  get newImage() {
+    return unmarshalItem(this.data.NewImage);
   }
 
-  get oldDocument() {
-    return new Document(this.data.OldDocument);
+  get oldImage() {
+    return unmarshalItem(this.data.OldImage);
   }
 
   get data() {