+++ src/events/lambdaEvent.js
@@ -1,3 +1,5 @@
+import { OK } from '../global';
+
 /**
  * A Lambda Callback signals to AWS Lambda that the function has completed
  * either successfully or otherwise. It can contain a result body (as a string)
@@ -16,6 +18,37 @@
  */
 export default class {
   /**
+   * A new instance of LambdaEvent
+   *
+   * @param {object} event - The event being received from AWS Lambda. The content of this object is
+   *  determined greatly by the stream it has originated from.
+   * @param {string} event.Records - The records being processed by this Lambda function.
+   * @param {object} context - The context of the event being received
+   *  [See here]{@link http://docs.aws.amazon.com/lambda/latest/dg/nodejs-prog-model-context.html}
+   * @param {lambdaCallback} cb - The callback function passed through from Lambda
+   *
+   * @see {@link http://docs.aws.amazon.com/lambda/latest/dg/eventsources.html#eventsources-ddb-update|Sample DynamoDB Event}
+   */
+  constructor(event, context, cb) {
+    /**
+     * The raw event received by Lambda
+     * @member {Object} LambdaEvent#event
+     */
+    this.event = event;
+    /**
+     * The raw context object received by Lambda
+     * @member {Object} LambdaEvent#context
+     */
+    this.context = context;
+    /**
+     * The response object allowing any extending class
+     * to respond to Cloudformation.
+     * @member {Function} LambdaEvent#cb
+     */
+    this.cb = cb;
+  }
+
+  /**
    * A Class-wrapper, allowing any class to handle and process events
    * in a similar manner, without making assumptions on how the function
    * will be executed.
@@ -30,5 +63,17 @@ export default class {
    */
   static wrap(Req, ...params) {
     return (ev, ctx, fn) => { new Req(ev, ctx, fn).perform(...params); };
+  }
+
+  perform() {
+    throw new Error('missing #perform implementation');
+  }
+
+  respond(status = OK, body = '') {
+    if (status === OK) {
+      return this.cb(null, body);
+    } else {
+      return this.cb(new Error(`[500] "${body}"`));
+    }
   }
 }