+++ src/index.js
@@ -1,12 +1,16 @@
 import { OK, ERROR } from './global';
 import Cloudformation from './events/Cloudformation';
 import SNS from './events/SNS';
+import Schedule from './events/Schedule';
 import DynamoDB from './events/DynamoDB';
+import CognitoStream from './events/CognitoStream';
 
 export {
   OK,
   ERROR,
   SNS,
+  Schedule,
   Cloudformation,
+  CognitoStream,
   DynamoDB,
 };