+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.1.2
+
+- Added CognitoStream event, to handle Cognito sync events streamed through Kinesis
+
+# 0.1.0
+
+- Added DynamoDB Events
+- Added SNS Events
+- Added Simple Schedule Event
+- Removed `this.response` and implemented `this.respond(OK | ERROR)`
+
 # 0.0.1
 
 - Initial Implementation